Add self-hosted queue fallback for Qstash-dependent features
- Create /simple endpoints for digest, summary, and digest-email that use
  internal API key auth instead of Qstash signature verification
- Add /api/cron/scheduled-actions endpoint to poll and execute due
  scheduled actions for self-hosted deployments
- Update scheduler to gracefully handle missing Qstash by setting
  schedulingStatus to CRON_FALLBACK instead of throwing
- Extract handlers from route files to enable code reuse between
  main endpoints and /simple fallback endpoints
- Add comment to clean endpoint noting Qstash requirement

Self-hosters can now run without Qstash by setting up a cron job
to poll /api/cron/scheduled-actions for delayed action processing.
