+++ apps/web/app/api/resend/summary/simple/route.ts
@@ -1,0 +1,22 @@
+import { NextResponse } from "next/server";
+import { headers } from "next/headers";
+import { withError } from "@/utils/middleware";
+import { handleSummaryRequest } from "@/app/api/resend/summary/handle-summary";
+import { env } from "@/env";
+import { isValidInternalApiKey } from "@/utils/internal-api";
+
+export const maxDuration = 60;
+
+// Fallback when Qstash is not in use
+export const POST = withError("resend/summary/simple", async (request) => {
+  if (env.QSTASH_TOKEN) {
+    return NextResponse.json({
+      error: "Qstash is set. This endpoint is disabled.",
+    });
+  }
+
+  if (!isValidInternalApiKey(await headers(), request.logger))
+    return NextResponse.json({ error: "Invalid API key" }, { status: 401 });
+
+  return handleSummaryRequest(request);
+});
